+++ PatternPractice/FactoryPattern/Goliath.swift
@@ -1,0 +1,15 @@
+//
+//  Goliath.swift
+//  PatternPractice
+//
+//  Created by jeon sangjun on 2017/10/13.
+//  Copyright © 2017年 jeon sangjun. All rights reserved.
+//
+
+import Foundation
+
+class Goliath: Unit {
+    func action() {
+        print("골리앗이 비행유닛을 쏩니다.")
+    }
+}
