+++ PatternPractice/FactoryPattern/Vulture.swift
@@ -1,0 +1,15 @@
+//
+//  Vulture.swift
+//  PatternPractice
+//
+//  Created by jeon sangjun on 2017/10/13.
+//  Copyright © 2017年 jeon sangjun. All rights reserved.
+//
+
+import Foundation
+
+class Vulture: Unit {
+    func action() {
+        print("벌처가 마인을 심습니다.")
+    }
+}
